+++ userspace/engine/lua/parser.lua
@@ -28,12 +28,6 @@
 
 local parser = {}
 
-parser.verbose = false
-
-function parser.set_verbose(verbose)
-   parser.verbose = verbose
-end
-
 local lpeg = require "lpeg"
 
 lpeg.locale(lpeg)
